automation/requirements.txt
yfinance
requests
duckduckgo-search
beautifulsoup4
python-dotenv
google-generativeai
curl_cffi
lxml
pydantic
